\name{ttTensor-package}
\alias{ttTensor-package}
\alias{ttTensor}
\docType{package}
\title{
\packageTitle{ttTensor}
}
\description{
\packageDescription{ttTensor}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ttTensor}
\packageIndices{ttTensor}
}
\author{
\packageAuthor{ttTensor}

Maintainer: \packageMaintainer{ttTensor}
}
\references{
I. V. Oseledets, (2011). Tensor-Train Decomposition. \emph{SIAM J. SCI. COMPUT.}

Yuan, Longhao, et. al., (2017). Completion of high order tensor data with missing entries via tensor-train decomposition. \emph{International Conference on Neural Information Processing}

I. V. Oseledets, et. al., (2010). TT-cross approximation for multidimensional arrays. \emph{Linear Algebra and its Applications}

Ali Civril, et. al., (2009). On selecting a maximum volume sub-matrix of a matrix and related problems. \emph{Theoretical Computer Science}
}

\keyword{ package }
\seealso{\code{\link{TTSVD}},\code{\link{TTWOPT}},\code{\link{TTCross}},\code{\link{skeleton.decomp}},\code{\link{maxvol}}}
\examples{
ls("package:ttTensor")
}